Move dataset resolver export functions to a separate file
