+++ lightly_studio/src/lightly_studio/resolvers/datasets_resolver/__init__.py
@@ -1,5 +1,9 @@
 """Resolvers for database operations."""
 
+from lightly_studio.resolvers.datasets_resolver.export import (
+    export,
+    get_filtered_samples_count,
+)
 from lightly_studio.resolvers.datasets_resolver.get_dataset_details import (
     get_dataset_details,
 )
@@ -8,6 +12,8 @@ from lightly_studio.resolvers.datasets_resolver.get_hierarchy import (
 )
 
 __all__ = [
+    "export",
     "get_dataset_details",
+    "get_filtered_samples_count",
     "get_hierarchy",
 ]
